devided functions to separate files
