+++ src/js/capitalize.js
@@ -1,0 +1,15 @@
+export default function capitalize(category) {
+  let categoryText = '';
+  switch (category) {
+    case 'idea':
+      categoryText = 'Idea';
+      break;
+    case 'thought':
+      categoryText = 'Random thought';
+      break;
+    case 'task':
+      categoryText = 'Task';
+      break;
+  }
+  return categoryText;
+}
